chore(email-verification):fix email verification routes [Fix #167730740]
